Connections to github isolated to its own class
